+++ showcase/shell-dashboard/src/lib/cell-model.ts
@@ -9,6 +9,16 @@
 
 import type { LiveStatusMap, StatusRow, State } from "./live-status";
 import { keyFor, CATALOG_TO_D5_KEY } from "./live-status";
+import { E2E_STALE_AFTER_MS, D4_STALE_AFTER_MS, isStale } from "./staleness";
+
+// Re-export the staleness windows so existing consumers that import them from
+// this module (e.g. `cell-model.test.ts`) keep resolving — the canonical
+// definitions now live in `./staleness`.
+export {
+  E2E_STALE_AFTER_MS,
+  D4_STALE_AFTER_MS,
+  LIVENESS_STALE_AFTER_MS,
+} from "./staleness";
 
 // ---------------------------------------------------------------------------
 // Types
@@ -16,39 +26,6 @@ import { keyFor, CATALOG_TO_D5_KEY } from "./live-status";
 
 export type TestStatus = "green" | "red" | "amber" | null;
 export type ChipColor = "green" | "amber" | "red" | "gray";
-
-/**
- * Staleness window for the `e2e:` dimension. The e2e-demos driver writes
- * `e2e:<slug>/<feature>` rows hourly (`schedule: "10 * * * *"`, see
- * harness/config/probes/e2e-demos.yml). When the driver stops writing
- * (a wedged browser pool, a dead probe pipeline), the last row freezes —
- * a green row then reads as a healthy D3 forever, masking the outage as a
- * false-green. Mirroring the original ">6h stale" model (see
- * live-status.ts), a green e2e row whose `observed_at` is older than this
- * window is downgraded to `degraded` (amber) so the staleness surfaces
- * instead of presenting as green. 6h tolerates several missed hourly ticks
- * before flagging, avoiding flapping on a single skipped run.
- */
-export const E2E_STALE_AFTER_MS = 6 * 60 * 60 * 1000;
-
-/**
- * Staleness window for the D4 (real-time chat/tools) dimension. The
- * `chat:<slug>`/`tools:<slug>` drivers write on their own cadence; a
- * frozen-green D4 row from a stalled driver must NOT credit D4 forever. A
- * green D4 row older than this window is downgraded to `degraded` (amber) so
- * the staleness surfaces, mirroring the D3/D5/D6 downgrade. Not env-tunable.
- */
-export const D4_STALE_AFTER_MS = 60 * 60 * 1000;
-
-/**
- * Staleness window for the D1/D2 (liveness — `health:<slug>`/`agent:<slug>`)
- * dimensions. Tighter than the e2e window because liveness probes are the
- * most frequently written signals; a frozen-green liveness row is a strong
- * stalled-driver indicator. Consumed by `depth-utils.ts` (cell-model has no
- * D1/D2 resolution — D3's failure implicitly covers the D1/D2 gate). Not
- * env-tunable.
- */
-export const LIVENESS_STALE_AFTER_MS = 45 * 60 * 1000;
 
 export interface TestLevel {
   exists: boolean;
@@ -238,17 +215,6 @@ function resolveD5(
     status: stateToTestStatus(worstState),
     row: worstRow,
   };
-}
-
-/**
- * Determine whether a row's `observed_at` is older than `maxAgeMs` relative
- * to `now`. An unparseable/missing timestamp is treated as NOT stale —
- * staleness must be a positive signal, never inferred from bad data.
- */
-function isStale(row: StatusRow, now: number, maxAgeMs: number): boolean {
-  const observedMs = Date.parse(row.observed_at);
-  if (Number.isNaN(observedMs)) return false;
-  return now - observedMs > maxAgeMs;
 }
 
 /**
